+++ src/bin/conserve.rs
@@ -72,6 +72,8 @@ enum Command {
         verbose: bool,
         #[structopt(long, short, number_of_values = 1)]
         exclude: Vec<String>,
+        #[structopt(long = "only", short = "i", number_of_values = 1)]
+        only_subtree: Option<Apath>,
         /// Restore the incomplete contents of an unfinished backup.
         #[structopt(long, requires = "backup")]
         incomplete: bool,
@@ -233,16 +235,20 @@ impl Command {
                 verbose,
                 force_overwrite,
                 exclude,
+                only_subtree,
                 incomplete,
             } => {
                 let band_selection = band_selection_policy_from_opt(backup, *incomplete);
                 let archive = Archive::open_path(archive)?;
+
                 let options = RestoreOptions {
                     print_filenames: *verbose,
                     excludes: excludes::from_strings(exclude)?,
+                    only_subtree: only_subtree.clone(),
                     band_selection,
                     overwrite: *force_overwrite,
                 };
+
                 let copy_stats = archive.restore(&destination, &options)?;
                 ui::println("Restore complete.");
                 copy_stats.summarize_restore(&mut stdout)?;